refactor: move parsing helpers to new module
